rag/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==0.11.0
chromadb==0.5.15
sentence-transformers==3.1.1
python-dotenv==1.0.1
pydantic==2.9.2